requirements.txt
django==3.0.5
djangorestframework==3.12.2
numpy==1.19.2
pandas==1.1.3
requests==2.24.0
pymongo==3.11.3
scipy==1.5.2
findpeaks==2.1.3